+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support `message_box` as an alias for `msg_box` (and `message_box_error` for `msg_box_error` too)
 - Tolerate `nil` input for any widget `String` attributes (e.g. `entry` `text` attribute)
 - Fix issue regarding arcs on Windows by auto-starting a figure if not started already (on Mac and Linux that is not needed)
+- Fix Color The Shapes (Circles) and rename back to Color The Circles due to fixing `circle` support on Windows
 
 ## 0.2.13